[feat:auth] prevent race by adding check:
Right now, if two people enter the server at the same time, there is a
chance of their playerdata getting mixed up, which opens a potential for
one of the players to be teleported to the other player's position.
Combat this by checking if the playerdata we received was actually for
the player that we are trying to auth at the moment.